change make to ./make and add ground to verbose debugging support
pegvm-verbose is a boolean that control if, when executing the code
will show the message "go inside number rule" when enter the number
rule code.

I don't know how to show this. I try using pegvm-current-rule, but
is a procedure. I see that format-id can transform from symbols to
identifiers, but how to transform to string, so we can display?